requirements.txt
streamlit>=1.28.0
tensorflow>=2.13.0
numpy>=1.24.0
Pillow>=9.5.0
six>=1.16.0
astunparse>=1.6.3
